Lots of JS development work
Each embedded SVG will now have a slider that allows the size of the
image to be adjusted down.  However, this isn't causing the Safari
browser to move the page break at all -- maybe the browser is stuck on
the original height of the HTMLElement and is doing the page break based
on that?  Nothing obvious turns up in the Inspector, though, so it's
going to be a bit tricky to figure it out.  Could just be a bug in
Safari, too -- I should try it in another browser (or two).